🐞 fix: 정렬기준 변경

+++ src/main/java/com/adoptpet/server/adopt/dto/response/ChatResponseDto.java
@@ -14,7 +14,7 @@ public class ChatResponseDto {
     private String type;
     private String content;
     private long sendDate;
-    private Integer readCount;
+    private long readCount;
 
     public ChatResponseDto(Chatting chatting) {
         this.id = chatting.getId();
